+++ src/hevm/src/EVM/SymExec.hs
@@ -34,6 +34,10 @@ import Data.Text (Text, splitOn, unpack)
 import qualified Control.Monad.State.Class as State
 import Control.Applicative
 
+data ProofResult a b c = Qed a | Cex b | Timeout c
+type VerifyResult = ProofResult (Tree BranchInfo) (Tree BranchInfo) (Tree BranchInfo)
+type EquivalenceResult = ProofResult ([VM], [VM]) VM ()
+
 -- | Convenience functions for generating large symbolic byte strings
 sbytes32, sbytes128, sbytes256, sbytes512, sbytes1024 :: Query ([SWord 8])
 sbytes32 = toBytes <$> freshVar_ @ (WordN 256)
@@ -149,14 +153,14 @@ data BranchInfo = BranchInfo
     _branchCondition    :: Maybe Whiff
   }
 
-doInterpret :: Fetch.Fetcher -> Maybe Integer -> VM -> Query (Tree BranchInfo)
-doInterpret fetcher maxIter vm = let
-      f (vm', cs) = Node (BranchInfo (if length cs == 0 then vm' else vm) Nothing) cs
-    in f <$> interpret' fetcher maxIter vm
+doInterpret :: Fetch.Fetcher -> Maybe Integer -> Maybe Integer -> VM -> Query (Tree BranchInfo)
+doInterpret fetcher maxIter askSmtIters vm = let
+      f (vm', cs) = Node (BranchInfo (if null cs then vm' else vm) Nothing) cs
+    in f <$> interpret' fetcher maxIter askSmtIters vm
 
-interpret' :: Fetch.Fetcher -> Maybe Integer -> VM -> Query (VM, [(Tree BranchInfo)])
-interpret' fetcher maxIter vm = let
-  cont s = interpret' fetcher maxIter $ execState s vm
+interpret' :: Fetch.Fetcher -> Maybe Integer -> Maybe Integer -> VM -> Query (VM, [Tree BranchInfo])
+interpret' fetcher maxIter askSmtIters vm = let
+  cont s = interpret' fetcher maxIter askSmtIters $ execState s vm
   in case view EVM.result vm of
 
     Nothing -> cont exec1
@@ -166,7 +170,7 @@ interpret' fetcher maxIter vm = let
       iteration = num $ fromMaybe 0 $ view (iterations . at codelocation) vm
       -- as an optimization, we skip consulting smt
       -- if we've been at the location less than 5 times
-      in if iteration < (max (fromMaybe 0 maxIter) 5)
+      in if iteration < (fromMaybe 5 askSmtIters)
          then cont $ continue EVM.Unknown
          else io (fetcher q) >>= cont
 
@@ -179,10 +183,10 @@ interpret' fetcher maxIter vm = let
           rvm = execState (continue False) vm
           in do
             push 1
-            (leftvm, left) <- interpret' fetcher maxIter lvm
+            (leftvm, left) <- interpret' fetcher maxIter askSmtIters lvm
             pop 1
             push 1
-            (rightvm, right) <- interpret' fetcher maxIter rvm
+            (rightvm, right) <- interpret' fetcher maxIter askSmtIters rvm
             pop 1
             return (vm, [Node (BranchInfo leftvm (Just whiff)) left, Node (BranchInfo rightvm (Just whiff)) right])
         Just n -> cont $ continue (not n)
@@ -195,10 +199,11 @@ interpret' fetcher maxIter vm = let
 -- | returns a list of possible final evm states
 interpret
   :: Fetch.Fetcher
-  -> Maybe Integer --max iterations
+  -> Maybe Integer -- max iterations
+  -> Maybe Integer -- ask smt iterations
   -> Stepper a
   -> StateT VM Query [a]
-interpret fetcher maxIter =
+interpret fetcher maxIter askSmtIters =
   eval . Operational.view
 
   where
@@ -212,47 +217,47 @@ interpret fetcher maxIter =
     eval (action Operational.:>>= k) =
       case action of
         Stepper.Exec ->
-          exec >>= interpret fetcher maxIter . k
+          exec >>= interpret fetcher maxIter askSmtIters . k
         Stepper.Run ->
-          run >>= interpret fetcher maxIter . k
+          run >>= interpret fetcher maxIter askSmtIters . k
         Stepper.IOAct q ->
-          mapStateT io q >>= interpret fetcher maxIter . k
+          mapStateT io q >>= interpret fetcher maxIter askSmtIters . k
         Stepper.Ask (EVM.PleaseChoosePath _ continue) -> do
           vm <- get
           case maxIterationsReached vm maxIter of
             Nothing -> do
               push 1
-              a <- interpret fetcher maxIter (Stepper.evm (continue True) >>= k)
+              a <- interpret fetcher maxIter askSmtIters (Stepper.evm (continue True) >>= k)
               put vm
               pop 1
               push 1
-              b <- interpret fetcher maxIter (Stepper.evm (continue False) >>= k)
+              b <- interpret fetcher maxIter askSmtIters (Stepper.evm (continue False) >>= k)
               pop 1
               return $ a <> b
             Just n ->
-              interpret fetcher maxIter (Stepper.evm (continue (not n)) >>= k)
+              interpret fetcher maxIter askSmtIters (Stepper.evm (continue (not n)) >>= k)
         Stepper.Wait q -> do
           let performQuery = do
                 m <- liftIO (fetcher q)
-                interpret fetcher maxIter (Stepper.evm m >>= k)
+                interpret fetcher maxIter askSmtIters (Stepper.evm m >>= k)
 
           case q of
             PleaseAskSMT _ _ continue -> do
               codelocation <- getCodeLocation <$> get
-              iteration <- num <$> fromMaybe 0 <$> use (iterations . at codelocation)
+              iteration <- num . fromMaybe 0 <$> use (iterations . at codelocation)
 
               -- if this is the first time we are branching at this point,
               -- explore both branches without consulting SMT.
               -- Exploring too many branches is a lot cheaper than
               -- consulting our SMT solver.
-              if iteration < (max (fromMaybe 0 maxIter) 5)
-              then interpret fetcher maxIter (Stepper.evm (continue EVM.Unknown) >>= k)
+              if iteration < (fromMaybe 5 askSmtIters)
+              then interpret fetcher maxIter askSmtIters (Stepper.evm (continue EVM.Unknown) >>= k)
               else performQuery
 
             _ -> performQuery
 
         Stepper.EVM m ->
-          State.state (runState m) >>= interpret fetcher maxIter . k
+          State.state (runState m) >>= interpret fetcher maxIter askSmtIters . k
 
 maxIterationsReached :: VM -> Maybe Integer -> Maybe Bool
 maxIterationsReached _ Nothing = Nothing
@@ -266,7 +271,7 @@ maxIterationsReached vm (Just maxIter) =
 type Precondition = VM -> SBool
 type Postcondition = (VM, VM) -> SBool
 
-checkAssert :: [Word256] -> ByteString -> Maybe (Text, [AbiType]) -> [String] -> Query (Either (Tree BranchInfo) (Tree BranchInfo), VM)
+checkAssert :: [Word256] -> ByteString -> Maybe (Text, [AbiType]) -> [String] -> Query (VerifyResult, VM)
 checkAssert errs c signature' concreteArgs = verifyContract c signature' concreteArgs SymbolicS (const sTrue) (Just $ checkAssertions errs)
 
 {- |Checks if an assertion violation has been encountered
@@ -305,12 +310,12 @@ allPanicCodes = [ 0x00, 0x01, 0x11, 0x12, 0x21, 0x22, 0x31, 0x32, 0x41, 0x51 ]
 panicMsg :: Word256 -> ByteString
 panicMsg err = (selector "Panic(uint256)") <> (encodeAbiValue $ AbiUInt 256 err)
 
-verifyContract :: ByteString -> Maybe (Text, [AbiType]) -> [String] -> StorageModel -> Precondition -> Maybe Postcondition -> Query (Either (Tree BranchInfo) (Tree BranchInfo), VM)
+verifyContract :: ByteString -> Maybe (Text, [AbiType]) -> [String] -> StorageModel -> Precondition -> Maybe Postcondition -> Query (VerifyResult, VM)
 verifyContract theCode signature' concreteArgs storagemodel pre maybepost = do
     preStateRaw <- abstractVM signature' concreteArgs theCode  storagemodel
     -- add the pre condition to the pathconditions to ensure that we are only exploring valid paths
     let preState = over constraints ((++) [(pre preStateRaw, Todo "assumptions" [])]) preStateRaw
-    v <- verify preState Nothing Nothing maybepost
+    v <- verify preState Nothing Nothing Nothing maybepost
     return (v, preState)
 
 pruneDeadPaths :: [VM] -> [VM]
@@ -347,36 +352,41 @@ leaves (Node x []) = [_vm x]
 leaves (Node _ xs) = concatMap leaves xs
 
 -- | Symbolically execute the VM and check all endstates against the postcondition, if available.
--- Returns `Right (Tree BranchInfo)` if the postcondition can be violated, or
--- or `Left (Tree BranchInfo)`, if the postcondition holds for all endstates.
-verify :: VM -> Maybe Integer -> Maybe (Fetch.BlockNumber, Text) -> Maybe Postcondition -> Query (Either (Tree BranchInfo) (Tree BranchInfo))
-verify preState maxIter rpcinfo maybepost = do
+verify :: VM -> Maybe Integer -> Maybe Integer -> Maybe (Fetch.BlockNumber, Text) -> Maybe Postcondition -> Query VerifyResult
+verify preState maxIter askSmtIters rpcinfo maybepost = do
   smtState <- queryState
-  tree <- doInterpret (Fetch.oracle (Just smtState) rpcinfo False) maxIter preState
+  tree <- doInterpret (Fetch.oracle (Just smtState) rpcinfo False) maxIter askSmtIters preState
   case maybepost of
     (Just post) -> do
       let livePaths = pruneDeadPaths $ leaves tree
-      -- can also do these queries individually (even concurrently!). Could save time and report multiple violations
+          -- have we hit max iterations at any point in a given path
+          maxReached :: VM -> Bool
+          maxReached p = case maxIter of
+            Just maxI -> any (>= (fromInteger maxI)) (view iterations p)
+            Nothing -> False
+          -- is there any path which can possibly violate the postcondition?
+          -- can also do these queries individually (even concurrently!). Could save time and report multiple violations
           postC = sOr $ fmap (\postState -> (sAnd (fst <$> view constraints postState)) .&& sNot (post (preState, postState))) livePaths
-      -- is there any path which can possibly violate
-      -- the postcondition?
       resetAssertions
       constrain postC
       io $ putStrLn "checking postcondition..."
       checkSat >>= \case
         Unk -> do io $ putStrLn "postcondition query timed out"
-                  return $ Left tree
-        Unsat -> do io $ putStrLn "Q.E.D."
-                    return $ Left tree
-        Sat -> return $ Right tree
+                  return $ Timeout tree
+        Unsat -> do
+          if any maxReached livePaths
+            then io $ putStrLn "WARNING: max iterations reached, execution halted prematurely"
+            else io $ putStrLn "Q.E.D."
+          return $ Qed tree
+        Sat -> return $ Cex tree
         DSat _ -> error "unexpected DSAT"
 
     Nothing -> do io $ putStrLn "Nothing to check"
-                  return $ Left tree
+                  return $ Qed tree
 
 -- | Compares two contract runtimes for trace equivalence by running two VMs and comparing the end states.
-equivalenceCheck :: ByteString -> ByteString -> Maybe Integer -> Maybe (Text, [AbiType]) -> Query (Either ([VM], [VM]) VM)
-equivalenceCheck bytecodeA bytecodeB maxiter signature' = do
+equivalenceCheck :: ByteString -> ByteString -> Maybe Integer -> Maybe Integer -> Maybe (Text, [AbiType]) -> Query EquivalenceResult
+equivalenceCheck bytecodeA bytecodeB maxiter askSmtIters signature' = do
   let
     bytecodeA' = if BS.null bytecodeA then BS.pack [0] else bytecodeA
     bytecodeB' = if BS.null bytecodeB then BS.pack [0] else bytecodeB
@@ -392,10 +402,10 @@ equivalenceCheck bytecodeA bytecodeB maxiter signature' = do
 
   smtState <- queryState
   push 1
-  aVMs <- doInterpret (Fetch.oracle (Just smtState) Nothing False) maxiter preStateA
+  aVMs <- doInterpret (Fetch.oracle (Just smtState) Nothing False) maxiter askSmtIters preStateA
   pop 1
   push 1
-  bVMs <- doInterpret (Fetch.oracle (Just smtState) Nothing False) maxiter preStateB
+  bVMs <- doInterpret (Fetch.oracle (Just smtState) Nothing False) maxiter askSmtIters preStateB
   pop 1
   -- Check each pair of endstates for equality:
   let differingEndStates = uncurry distinct <$> [(a,b) | a <- pruneDeadPaths (leaves aVMs), b <- pruneDeadPaths (leaves bVMs)]
@@ -428,9 +438,9 @@ equivalenceCheck bytecodeA bytecodeB maxiter signature' = do
   constrain $ sOr differingEndStates
 
   checkSat >>= \case
-     Unk -> error "solver said unknown!"
-     Sat -> return $ Right preStateA
-     Unsat -> return $ Left (leaves aVMs, leaves bVMs)
+     Unk -> return $ Timeout ()
+     Sat -> return $ Cex preStateA
+     Unsat -> return $ Qed (leaves aVMs, leaves bVMs)
      DSat _ -> error "unexpected DSAT"
 
 both' :: (a -> b) -> (a, a) -> (b, b)